+++ webapp/src/components/rows/EditRow.js
@@ -1,37 +1,34 @@
 import React, {useState} from 'react'
-// import { useMutation } from '@apollo/client';
-// import { EDIT_TRANSACTION } from '../../gql/mutation.js'
+import { useMutation } from '@apollo/client';
+import { EDIT_TRANSACTION } from '../../gql/mutation.js'
 
-const EditRow = ({
-    editFormData, 
-    handleEditFormChange, 
-    handleCancel,
-    setCredit,
-    setDebit
-    }) => {
+const EditRow = ({editFormData, handleEditFormChange, handleCancel}) => {
 
-    // const [ editTransaction, {error} ] = useMutation(EDIT_TRANSACTION);
+    const [ editTransaction, {error} ] = useMutation(EDIT_TRANSACTION);
 
-    // const [debit, setDebit] = useState(editFormData.debit);
-    // const [credit, setCredit] = useState(editFormData.credit);
+    const [debit, setDebit] = useState(editFormData.debit);
+    const [credit, setCredit] = useState(editFormData.credit);
 
-    // const editTrans = (e) =>{
-    //     e.preventDefault()
-    //     editTransaction({
-    //         variables: {
-    //             id: editFormData.id,
-    //             user_id: editFormData.user_id,
-    //             description: editFormData.description,
-    //             merchant_id: editFormData.merchant_id,
-    //             debit: debit,
-    //             credit: credit,
-    //             amount: parseInt(editFormData.amount)
-    //         }
-    //     });
-    //     if (error){
-    //         console.log(error)
-    //     }
-    // }
+    console.log(editFormData)
+
+    const editTrans = (e) =>{
+        e.preventDefault()
+        editTransaction({
+            variables: {
+                id: editFormData.id,
+                user_id: editFormData.user_id,
+                description: editFormData.description,
+                merchant_id: editFormData.merchant_id,
+                debit: debit,
+                credit: credit,
+                amount: parseInt(editFormData.amount)
+            }
+        });
+        if (error){
+            console.log(error)
+        }
+        // location.reload()
+    }
 
 
 
@@ -97,7 +94,7 @@ const EditRow = ({
         >
         </input></td>
         <td>
-            <button type='submit'>Save</button>
+            <button type='submit' onClick={editTrans}>Save</button>
             <button type='button' onClick={handleCancel}>Cancel</button>
         </td>
         </tr>
