--- webapp/src/home/index.js
@@ -1,1 +1,0 @@
-export { Home } from './home-page'
